requirements.txt
anyio==3.7.0
asgiref==3.7.2
attrs==22.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
future==0.18.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
Naked==0.1.32
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.12.0
pytest==7.2.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
status==0.2.5
stripe==5.4.0
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
